+++ src/components/Div.tsx
@@ -1,13 +1,12 @@
-import type { CSSProperties, FC, ReactNode } from 'react';
-import { createReactStyles } from '../helpers';
+import type { ParentComponent } from 'solid-js';
+import { createStyles, type CSSProperties } from '../helpers';
 
 export type DivProps = {
   style?: CSSProperties;
-  children?: ReactNode;
 };
 
-const Div: FC<DivProps> = ({ children, style }) => {
-  return <div style={createReactStyles(style)}>{children}</div>;
-};
+const Div: ParentComponent<DivProps> = (props) => (
+  <div style={createStyles(props.style)}>{props.children}</div>
+);
 
 export default Div;