+++ src/components/Paragraph.tsx
@@ -1,13 +1,12 @@
-import type { CSSProperties, FC, ReactNode } from 'react';
-import { createReactStyles } from '../helpers';
+import type { ParentComponent } from 'solid-js';
+import { createStyles, type CSSProperties } from '../helpers';
 
 export type ParagraphProps = {
   style?: CSSProperties;
-  children?: ReactNode;
 };
 
-const Paragraph: FC<ParagraphProps> = ({ children, style }) => {
-  return <p style={createReactStyles(style)}>{children}</p>;
-};
+const Paragraph: ParentComponent<ParagraphProps> = (props) => (
+  <p style={createStyles(props.style)}>{props.children}</p>
+);
 
 export default Paragraph;